Updated yaml-adapter-tests
