+++ test/scripts/adapters/index.js
@@ -3,4 +3,5 @@
 describe('Adapters', () => {
   require('./json');
   require('./xml');
+  require('./yaml');
 });
