+++ test/scripts/adapters/yaml/async/index.js
@@ -1,0 +1,8 @@
+// jshint esversion: 8
+
+describe('YamlFileAdapter', () => {
+
+  require('./save');
+  require('./load');
+
+});
