+++ test/scripts/adapters/yaml/index.js
@@ -1,0 +1,6 @@
+// jshint esversion: 8
+
+describe('yaml', () => {
+  require('./async');
+  require('./sync');
+});
