+++ test/scripts/adapters/yaml/sync/index.js
@@ -1,0 +1,8 @@
+// jshint esversion: 8
+
+describe('SyncYamlFileAdapter', () => {
+
+  require('./save');
+  require('./load');
+
+});
